Wire up announcement cancel interactions
